Add CLI, ensure compatibility as external mount helper
This CLI added is in accordance with the spec for external mount helpers
as described in mount(8). This means that we can distribute this
application to `/usr/sbin/mount.s3wofs` and/or
`/usr/sbin/mount.fuse.s3wofs` and use the mount-command to mount our
filesystem.

This in turn means that everything that supports regular mounts (like
systemd mount and automount units), will support this filesystem.