Add support to daemonize on call
For our binary to be a fully compatible "external mount helper", it
needs to perform the mount and return afterwards. To achieve this, we
now daemonize unless foreground execution is requested, which means the
initial call to the binary will return after the initial setup and mount
the filesystem detached from the session that asked for the mount.

To still have logs in the detached state, the logger is reconfigured to
push its logs to journald.